Split data specific retriever helpers into separate "_mixins.py".
