+++ src/hawk/scout/retrieval/retriever.py
@@ -10,11 +10,7 @@ import threading
 import time
 from abc import ABCMeta, abstractmethod
 from dataclasses import dataclass
-from io import BytesIO
 from pathlib import Path
-
-from logzero import logger
-from PIL import Image
 
 from ...classes import (
     NEGATIVE_CLASS,
@@ -38,8 +34,6 @@ from ..stats import (
     collect_metrics_total,
 )
 
-THUMBNAIL_SIZE = (256, 256)
-
 
 @dataclass
 class RetrieverStats:
@@ -61,6 +55,10 @@ class RetrieverBase(metaclass=ABCMeta):
 
     @abstractmethod
     def put_objectid(self, object_id: ObjectId) -> None:
+        pass
+
+    @abstractmethod
+    def get_objectid(self, timeout: float | None = None) -> ObjectId | None:
         pass
 
     @abstractmethod
@@ -185,7 +183,7 @@ class Retriever(RetrieverBase):
             self.queue_length.dec()
             self.dropped_objects.inc()
 
-    def _get_objectid(self, timeout: float | None = None) -> ObjectId | None:
+    def get_objectid(self, timeout: float | None = None) -> ObjectId | None:
         """Pop a single ML ready object identifier from the result queue."""
         try:
             object_id = self.result_queue.get(timeout=timeout)
@@ -236,7 +234,7 @@ class Retriever(RetrieverBase):
                 if wait_time < 0:
                     break
 
-            object_id = self._get_objectid(timeout=wait_time)
+            object_id = self.get_objectid(timeout=wait_time)
             if object_id is None:
                 break
 
@@ -249,74 +247,3 @@ class Retriever(RetrieverBase):
             time_now = time.time()
 
         return (object_ids, hawk_objects)
-
-
-class ThumbnailImageMixin(RetrieverBase):
-    """Uses Pillow to derive a thumbnail image from the ML ready data."""
-
-    def get_oracle_data(self, object_id: ObjectId) -> list[HawkObject]:
-        ml_object = self.get_ml_data(object_id)
-        if ml_object is None or not ml_object.media_type.startswith("image/"):
-            raise ValueError("Generic get_oracle_data only works for images")
-
-        with BytesIO(ml_object.content) as f:
-            image = Image.open(f)
-
-        image = image.convert("RGB")
-
-        # crop to centered square
-        if image.size[0] != image.size[1]:
-            short_edge = min(image.size)
-            left = (image.size[0] - short_edge) // 2
-            top = (image.size[1] - short_edge) // 2
-            right = left + short_edge
-            bottom = top + short_edge
-            image = image.crop((left, top, right, bottom))
-
-        # resize to THUMBNAIL_SIZE
-        # image.thumbnail(THUMBNAIL_SIZE)
-        image = image.resize(THUMBNAIL_SIZE)
-
-        with BytesIO() as tmpfile:
-            image.save(tmpfile, format="JPEG", quality=85)
-            content = tmpfile.getvalue()
-
-        return [HawkObject(content=content, media_type="image/jpeg")]
-
-
-class LegacyRetrieverMixin(ThumbnailImageMixin):
-    """Get tile and groundtruth based on the legacy ObjectId format."""
-
-    def __init__(self) -> None:
-        super().__init__()
-
-        # TODO: we probably should be more restrictive about where images
-        # are allowed to be retrieved from.
-        # Random retriever sets this to the parent of the INDEXES directory.
-        self._data_root = Path("/")
-
-    def get_ml_data(self, object_id: ObjectId) -> HawkObject | None:
-        """Return ML ready tile for inferencing or training."""
-        object_path = object_id._file_path(self._data_root)
-        if object_path is None:
-            logger.error(f"Unable to get path for {object_id}")
-            return None
-        try:
-            return HawkObject.from_file(object_path)
-        except FileNotFoundError:
-            logger.error(f"Unable to read {object_id}")
-            return None
-
-    def get_groundtruth(self, object_id: ObjectId) -> list[BoundingBox]:
-        """Return groundtruth for logging, statistics and scriptlabeler."""
-        # only handles classification groundtruth as it assumes it is stashed
-        # in the object id.
-        class_name = object_id._groundtruth()
-        if class_name is None:
-            return []
-
-        return [
-            BoundingBox(
-                x=0.5, y=0.5, w=1.0, h=1.0, class_name=class_name, confidence=1.0
-            )
-        ]
