+++ src/hawk/scout/retrieval/tile_retriever.py
@@ -13,7 +13,8 @@ from ...classes import ClassLabel
 from ...objectid import ObjectId
 from ...proto.messages_pb2 import FileDataset
 from ..stats import collect_metrics_total
-from .retriever import LegacyRetrieverMixin, Retriever
+from .retriever import Retriever
+from .retriever_mixins import LegacyRetrieverMixin
 
 
 class TileRetriever(Retriever, LegacyRetrieverMixin):
